use shadow map

+++ app/src/main/assets/shaders/GLES/shadowMap/Animation.frag
@@ -1,6 +1,6 @@
 #version 300 es
 
-precision highp float; // highp mediump lowp
+precision lowp float; // highp mediump lowp
 
 void main()
 {
